refactor: 빈 생성 책임 변경
